requirements.txt
azure-common==1.1.16
azure-nspkg==3.0.2
azure-servicebus==0.21.1
certifi==2018.8.24
chardet==3.0.4
idna==2.7
python-dotenv==0.9.1
requests==2.19.1
urllib3==1.23
